reverting to repo before making new branch

+++ src/buffers/heaps.c
@@ -1,5 +1,6 @@
-#include "global.h"
+#include <ultra64.h>
+#include <global.h>
 
 u8 gAudioHeap[0x138000];
 
-u8 gSystemHeap[UNK_SIZE];
+u32 gSystemHeap[UNK_SIZE];
